Fixed logging and regex regression in 3.11
