+++ .pre-commit-config.yaml
@@ -59,3 +59,5 @@ repos:
     rev: 0.27.1
     hooks:
       - id: check-azure-pipelines
+ci:
+  autofix_prs: false
